requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via uvicorn
distro==1.9.0
    # via openai
fastapi==0.115.6
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
jiter==0.13.0
    # via openai
openai==1.60.2
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.3
    # via uvicorn
sniffio==1.3.1
    # via openai
starlette==0.41.3
    # via fastapi
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn[standard]==0.34.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
sqlalchemy
psycopg[binary]
pgvector
requests
python-multipart
pypdf
xxhash